requirements.txt
fastapi
httpx
uvicorn
pydantic
python-dotenv
yt-dlp
gdown
openai-whisper
torch
torchaudio
python-multipart
